Rename admin to admin*
